pyproject.toml
[project]
name = "coder-sidekick"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "gradio>=6.2.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.1.6",
    "langgraph-checkpoint-sqlite>=3.0.1",
    "openai>=2.13.0",
    "pydantic>=2.12.5",
    "sendgrid>=6.12.5",
    "wikipedia>=1.4.0",
]
